Feat: 공감데시벨 계산 로직 추가

+++ src/main/java/org/quack/QUACKServer/service/UserService.java
@@ -109,9 +109,10 @@ public class UserService {
     public MyPageInfoResponse getMyPage(Long userId) {
         User user = getUserOrException(userId);
 
+        // 추후 DTO 프로젝션, 직접 쿼리 작성으로 성능 개선
         int reviewCount = reviewService.getReviewCountByUserId(userId);
         int savedRestaurantCount = savedRestaurantService.getSavedRestaurantCountByUserId(userId);
-        int empathyDecibel = reviewService.getEmpathyDecibelByUserId(userId);
+        double empathyDecibel = reviewService.getEmpathyDecibelByUserId(userId);
 
         return MyPageInfoResponse.of(
                 user.getNickname(),
